+++ packages/ice/src/middlewares/ssr/renderMiddleware.ts
@@ -64,6 +64,7 @@ export default function createRenderMiddleware(options: Options): Middleware {
         req,
         res,
       };
+
       serverModule.renderToResponse(requestContext, {
         renderMode,
         documentOnly,