+++ packages/ice/src/service/serverCompiler.ts
@@ -139,12 +139,6 @@ export function createServerCompiler(options: Options) {
     for (const [key, value] of Object.entries(expandedEnvs)) {
       define[`import.meta.env.${key}`] = JSON.stringify(value);
     }
-    // Add process.env.
-    Object.keys(process.env)
-      .filter((key) => /^ICE_/.test(key) || key === 'NODE_ENV')
-      .forEach((key) => {
-        define[`import.meta.env.${key}`] = JSON.stringify(process.env[key]);
-      });
 
     const format = customBuildOptions?.format || 'esm';
 
@@ -231,7 +225,7 @@ export function createServerCompiler(options: Options) {
       };
     } catch (error) {
       logger.error(
-        'Server compile error.',
+        'Server compiled with errors.',
         `\nEntryPoints: ${JSON.stringify(buildOptions.entryPoints)}`,
         `\n${error.message}`,
       );