+++ packages/ice/src/webpack/DataLoaderPlugin.ts
@@ -4,7 +4,7 @@ import type { Compiler } from 'webpack';
 import webpack from '@ice/bundles/compiled/webpack/index.js';
 import type { Context } from 'build-scripts';
 import type { ServerCompiler, PluginData } from '../types/plugin.js';
-import { RUNTIME_TMP_DIR } from '../constant.js';
+import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR } from '../constant.js';
 import { getRoutePathsFromCache } from '../utils/getRoutePaths.js';
 
 const pluginName = 'DataLoaderPlugin';
@@ -13,20 +13,23 @@ const { RawSource } = webpack.sources;
 export default class DataLoaderPlugin {
   private serverCompiler: ServerCompiler;
   private rootDir: string;
+  private target: string;
   private dataCache: Map<string, string>;
   private getAllPlugin: Context['getAllPlugin'];
 
   public constructor(options: {
     serverCompiler: ServerCompiler;
     rootDir: string;
+    target: string;
     dataCache: Map<string, string>;
     getAllPlugin?: Context['getAllPlugin'];
   }) {
-    const { serverCompiler, rootDir, dataCache, getAllPlugin } = options;
+    const { serverCompiler, rootDir, dataCache, getAllPlugin, target } = options;
     this.serverCompiler = serverCompiler;
     this.rootDir = rootDir;
     this.dataCache = dataCache;
     this.getAllPlugin = getAllPlugin;
+    this.target = target;
   }
 
   public apply(compiler: Compiler) {
@@ -49,8 +52,7 @@ export default class DataLoaderPlugin {
         if (fse.existsSync(filePath)) {
           const { outputFiles, error } = await this.serverCompiler(
             {
-              // Code will be transformed by @swc/core reset target to esnext make modern js syntax do not transformed.
-              target: 'esnext',
+              target: 'es6', // should not set to esnext, https://github.com/alibaba/ice/issues/5830
               entryPoints: [filePath],
               write: false,
               logLevel: 'silent', // The main server compile process will log it.
@@ -58,10 +60,14 @@ export default class DataLoaderPlugin {
             {
               swc: {
                 keepExports,
-                keepPlatform: 'web',
                 getRoutePaths: () => {
                   return getRoutePathsFromCache(this.dataCache);
                 },
+              },
+              runtimeDefineVars: {
+                [IMPORT_META_TARGET]: JSON.stringify(this.target),
+                // `data-loader.js` runs in the browser.
+                [IMPORT_META_RENDERER]: JSON.stringify('client'),
               },
               preBundle: false,
               externalDependencies: false,