+++ packages/runtime/src/index.ts
@@ -15,6 +15,7 @@ import type {
   AppProvider,
   RouteWrapper,
   RenderMode,
+  DistType,
   DataLoaderConfig,
   RouteWrapperConfig,
 } from './types.js';
@@ -41,6 +42,7 @@ import { routerHistory as history } from './history.js';
 import KeepAliveOutlet from './KeepAliveOutlet.js';
 import ClientOnly from './ClientOnly.js';
 import useMounted from './useMounted.js';
+import { withSuspense, useSuspenseData } from './Suspense.js';
 
 export {
   getAppConfig,
@@ -80,6 +82,9 @@ export {
   AppErrorBoundary,
   ClientOnly,
   useMounted,
+
+  withSuspense,
+  useSuspenseData,
 };
 
 export type {
@@ -93,6 +98,7 @@ export type {
   AppProvider,
   RouteWrapper,
   RenderMode,
+  DistType,
   DataLoaderConfig,
   RunClientAppOptions,
 };